Decouple place order modal from product page, closes #329
